[FIX] website: show megamenu bottom in scrolled mobile
Scenario: add a mega menu, in mobile scroll to not be at top of page,
and click to open hamburger menu then click to open mega menu content.

Result: the mega menu is mostly shown thanks to
579160b08c6e73d809be79f561214118f1284532 but a small part at the bottom
is inaccessible.

Reason: the height needs to take into account the margin bottom so it
can be shown fully on the page.

Fix: add a calc to remove the height of the margin-top.

opw-4737577
opw-4737823
opw-4738673

closes odoo/odoo#207013

Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>